Created some functions in the scritp.js file.
Created the 'playerSelection' function which prompts the user to enter
an option between 'rock', 'paper', 'scissors' and transforms it
into lower case.
Returns an 'error' message in case the entered input is not one
of the available options.

Created the function 'computerSelection' which makes the 'computer player'
randomly pick an option between 'rock', 'paper' and 'scissors'.

Created the function 'singleRound' which takes two arguments, the
'playerSelection' function and the 'computerSelection' function.
It compares both choices, the players choice and the computers choice,
and returns a string declaring if the user lost, won, or if it was a tie.